pyproject.toml
[tool.poetry]
name = "p-web-2-11"
version = "0.1.0"
description = ""
authors = ["Hlushchenko"]
readme = "README.md"
packages = [{include = "p"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
sqlalchemy = "^2.0.3"
psycopg2-binary = "^2.9.5"
alembic = "^1.9.4"
pydantic = {extras = ["dotenv"], version = "^1.10.5"}
libgravatar = "^1.0.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
fastapi-mail = "^1.2.6"
fastapi-limiter = "^0.1.5"
python-dotenv = "^1.0.0"
cloudinary = "^1.32.0"
sphinx = "^6.1.3"
pytest = "^7.2.2"


[tool.poetry.group.dev.dependencies]
sphinx = "^6.1.3"


[tool.poetry.group.test.dependencies]
httpx = "^0.23.3"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
